initial changes for angular 7 upgrade
