+++ src/app/app.component.ts
@@ -4,7 +4,7 @@ import { MultiselectComponent } from './multiselect/multiselect.component';
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
-  styleUrls: ['./app.component.css']
+  styleUrls: ['./app.component.scss']
 })
 export class AppComponent implements OnInit {
   title = 'Dataverse Report';
