+++ src/tsconfig.app.json
@@ -2,9 +2,7 @@
   "extends": "../tsconfig.json",
   "compilerOptions": {
     "outDir": "../out-tsc/app",
-    "baseUrl": "./",
-    "module": "es2015",
-    "types": ["d3"]
+    "types": []
   },
   "exclude": [
     "test.ts",
